refactor(search): one module per table behind the Store class
`store.ts` was 2460 lines and the worst maintainability score in the
repository (MI -84.1) - schema lifecycle, the writer lock, the KV state
table, documents, chunks, vectors, the FTS lane, trigrams, links,
aliases, typed relations, tiers, entities and the query-cache table, all
as methods on one class.

The class is the abstraction 25 modules hold, so it does not move: its
full public surface is unchanged and the methods became thin delegations
to seventeen modules that take the database handle as an explicit
parameter. Every module is named for the table or lifecycle concern it
owns; the largest is 417 lines.

The on-disk format is untouched: `schema.ts` has an empty diff, the
schema version is still 10, and no `CREATE TABLE` or `ALTER TABLE` text
exists anywhere in the split. An index built before this change opens
after it.

Six defects fixed while moving. The one that mattered: `trigramCandidates`
caught every database error and returned an empty candidate list, so an
unreadable index silently narrowed recall while looking healthy - the
documented intent was only "the trigram table is absent", and that is
now the only condition it stays quiet for; anything else raises
INDEX_UNREADABLE. Also a relation-endpoint SQL ladder copy-pasted into
three queries, two spellings of the same state read, index-revision
parsing duplicated so a malformed cell could mean two different corpus
generations, a ternary with identical arms, and a doc comment that had
drifted onto the wrong method.

The best-effort paths that are deliberate keep their behaviour and their
comments.
